requirements.txt
flask
requests
pyyaml
colorama
